refactor(project-tools): split init plan presentation helpers
