+++ .sampo/changesets/772-split-init-plan-presentation.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split retrofit init plan presentation helpers out of the main planning module.
